first working version with followers, OW_GFX_COMPRESS not functional

+++ src/fldeff_rocksmash.c
@@ -23,7 +23,7 @@ static void StartRockSmashFieldEffect(void);
 
 EWRAM_DATA struct MapPosition gPlayerFacingPosition = {};
 
-bool8 CheckObjectGraphicsInFrontOfPlayer(u8 graphicsId)
+bool8 CheckObjectGraphicsInFrontOfPlayer(u16 graphicsId)
 {
     u8 mapObjId;
